add option --e for except files translate and --f for specified files translate

+++ src/Console/InstallCommand.php
@@ -13,7 +13,7 @@ class InstallCommand extends Command
      *
      * @var string
      */
-    protected  $signature = 'atpro:translate';
+    protected  $signature = 'atpro:translate {--e=} {--f=}';
 
     /**
      * The console command description.
@@ -43,11 +43,24 @@ class InstallCommand extends Command
         $input['from'] = $this->ask('Your started language ?');
         $input['to'] = $this->ask('Yours translated list language seperated with commas (,) example: fr,es ... ?');
         $to = explode(',',$input['to']);
+        $except = $this->option('e') !== null ? explode(',',$this->option('e')) : [];
+        $files = $this->option('f') !== null ? explode(',',$this->option('f')) : [];
         $atproTranslateService = new AtproTranslateService();
         $this->info('Loading ...');
-        $atproTranslateService->translate($input['from'],$to);
-        $this->info('Files translated added successfully');
-        $atproTranslateService->deleteDirectory();
+        if (!empty($files) and !empty($except)) {
+            $this->error('Cant not combine files (--f) and except (--e) .');
+        }else{
+            if (!empty($except)){
+                $atproTranslateService->translateExept($input['from'],$to, $files);
+            }
+            else if (!empty($files)){
+                $atproTranslateService->translateList($input['from'],$to, $files);
+            }else{
+                $atproTranslateService->translate($input['from'],$to);
+            }
+            $this->info('Files translated added successfully');
+            $atproTranslateService->deleteDirectory();
+        }
 
     }
 }
